# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "emoji API",
    "description": "emoji API",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "local development"
    }
  ],
  "paths": {
    "/emoji/translate": {
      "post": {
        "operationId": "translate",
        "description": ":\u003calias\u003e:のような表現を含んだ文字列をemojiを使った文字列に変換する",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  }
                },
                "required": [
                  "text"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "default error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "emoji"
        ]
      }
    },
    "/emoji/suggest": {
      "post": {
        "operationId": "suggest",
        "description": "先頭一致で対応する文字列を探す",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prefix": {
                    "type": "string"
                  },
                  "sort": {
                    "type": "string",
                    "enum": [
                      "asc",
                      "desc"
                    ],
                    "default": "asc"
                  },
                  "limit": {
                    "type": "integer"
                  }
                },
                "required": [
                  "prefix",
                  "sort"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmojiDefinition"
                  }
                }
              }
            }
          },
          "default": {
            "description": "default error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "emoji"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "default error",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": false
      },
      "EmojiDefinition": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": ":dizzy:"
          },
          "char": {
            "type": "string",
            "example": "💫"
          }
        },
        "required": [
          "alias",
          "char"
        ],
        "additionalProperties": false
      }
    }
  }
}